+++ config/utils/visual-server.js
@@ -38,9 +38,9 @@
 
   function startCI() {
     return new Promise(function (resolve, reject) {
-      bsLocal = new browserstack.Local();
+      //bsLocal = new browserstack.Local();
       return server.listen(webpackCompiler.options.devServer.port, function () {
-        return bsLocal.start({
+        /*return bsLocal.start({
           key: process.env.BROWSER_STACK_ACCESS_KEY,
           onlyAutomate: true,
           forcelocal: true,
@@ -54,7 +54,9 @@
           } else {
             resolve();
           }
-        });
+        });*/
+        console.log('starting server');
+        return resolve();
       });
     });
   }
@@ -72,7 +74,7 @@
   function stopCI(exitCode) {
     server.close();
     rimraf.sync('webdriver-screenshots*/**/*+(full|regression).png', {});
-    if (bsLocal.isRunning()) {
+    if (bsLocal && bsLocal.isRunning()) {
       bsLocal.stop();
     }
 
@@ -80,7 +82,7 @@
 
   process.on('SIGINT', function () {
     stop();
-    if (bsLocal.isRunning()) {
+    if (bsLocal && bsLocal.isRunning()) {
       bsLocal.stop();
     }