+++ src/modules/modal/modal-close-args.ts
@@ -1,0 +1,4 @@
+export class SkyModalCloseArgs {
+  public reason: string;
+  public data: any;
+}