+++ src/modules/tabs/fixtures/tabs-fixtures.module.ts
@@ -1,19 +1,30 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { SkyModalModule } from '../../modal';
 
 import { SkyTabsModule } from '../';
 import { TabsetTestComponent } from './tabset.component.fixture';
+import { SkyWizardTestFormComponent } from './tabset-wizard.component.fixture';
 
 @NgModule({
   declarations: [
-    TabsetTestComponent
+    TabsetTestComponent,
+    SkyWizardTestFormComponent
   ],
   imports: [
     CommonModule,
+    FormsModule,
+    SkyModalModule,
     SkyTabsModule
   ],
   exports: [
-    TabsetTestComponent
+    TabsetTestComponent,
+    SkyWizardTestFormComponent
+  ],
+  entryComponents: [
+    SkyWizardTestFormComponent
   ]
 })
 export class SkyTabsFixturesModule { }