+++ src/modules/tabs/tabset.component.ts
@@ -6,8 +6,11 @@ import {
   DoCheck,
   ElementRef,
   EventEmitter,
+  Input,
+  OnDestroy,
   Output,
-  QueryList
+  QueryList,
+  ChangeDetectorRef
 } from '@angular/core';
 
 import { SkyTabComponent } from './tab.component';
@@ -20,7 +23,10 @@ import { SkyTabsetService } from './tabset.service';
   templateUrl: './tabset.component.html',
   providers: [SkyTabsetAdapterService, SkyTabsetService]
 })
-export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCheck {
+export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCheck, OnDestroy {
+  @Input()
+  public tabStyle = 'tabs';
+
   @Output()
   public newTab = new EventEmitter<any>();
 
@@ -32,13 +38,16 @@ export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCh
   @ContentChildren(SkyTabComponent)
   public tabs: QueryList<SkyTabComponent>;
 
+  private isDestroyed: boolean;
+
   constructor(
     private tabsetService: SkyTabsetService,
     private adapterService: SkyTabsetAdapterService,
-    private elRef: ElementRef
+    private elRef: ElementRef,
+    private changeRef: ChangeDetectorRef
   ) {
     tabsetService.tabDestroy.subscribe((tab: SkyTabComponent) => {
-      if (tab.active) {
+      if (!this.isDestroyed && tab.active) {
         let tabs = this.tabs.toArray();
         let tabIndex = tabs.indexOf(tab);
 
@@ -55,11 +64,15 @@ export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCh
   }
 
   public selectTab(tab: SkyTabComponent) {
-    this.tabs.forEach((existingTab) => {
+    let tabs = this.tabs.toArray();
+
+    for (let i = 0, n = tabs.length; i < n; i++) {
+      let existingTab = tabs[i];
+
       if (tab !== existingTab) {
         existingTab.active = false;
       }
-    });
+    }
 
     tab.active = true;
   }
@@ -94,11 +107,13 @@ export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCh
     this.adapterService.init(this.elRef);
 
     this.adapterService.overflowChange.subscribe((currentOverflow: boolean) => {
-      this.updateDisplayMode();
+      this.updateDisplayMode(currentOverflow);
     });
 
     setTimeout(() => {
-      this.updateDisplayMode();
+      this.adapterService.detectOverflow();
+      this.updateDisplayMode(this.adapterService.currentOverflow);
+      this.changeRef.detectChanges();
     }, 0);
   }
 
@@ -106,7 +121,12 @@ export class SkyTabsetComponent implements AfterContentInit, AfterViewInit, DoCh
     this.adapterService.detectOverflow();
   }
 
-  private updateDisplayMode() {
-    this.tabDisplayMode = this.adapterService.currentOverflow ? 'dropdown' : 'tabs';
+  public ngOnDestroy() {
+    this.tabsetService.destroy();
+    this.isDestroyed = true;
+  }
+
+  private updateDisplayMode(currentOverflow: boolean) {
+    this.tabDisplayMode = currentOverflow ? 'dropdown' : 'tabs';
   }
 }