+++ src/modules/wait/wait-page.component.ts
@@ -1,0 +1,17 @@
+import {
+  Component,
+  Input
+} from '@angular/core';
+
+@Component({
+  selector: 'sky-wait-page',
+  templateUrl: './wait-page.component.html'
+})
+export class SkyWaitPageComponent {
+
+  @Input()
+  public hasBlockingWait: boolean;
+
+  @Input()
+  public hasNonBlockingWait: boolean;
+}